Convert CanvasFactory into an interface
